Refactored index.py validation code into validation.py
